==> notebooks/part_2-2_spacy_multi_cat_test.ipynb <==
# Detecting and Classifying Toxic Comments
# Part 2-3: spaCy Multi-Category Test

Testing multiple, non-mutually exclusive categories with 1 instead of True

# Setup

## Python Library Imports


Resources:
- [pool]()

import pandas as pd
import numpy as np

from collections import Counter
import re
import random

# scikit learn imports
from sklearn.model_selection import train_test_split
from sklearn.metrics import classification_report

# # tqdm & time
# from tqdm.auto import tqdm
# import time
from timeit import default_timer as timer

%load_ext autoreload
%autoreload 2

## spaCy Setup & Imports

As mentioned previously, we'll be using spaCy version 2.3.5

# # check version
! python -m spacy info

# spaCy Imports
import spacy

from spacy.lang.en import English
spacy_stopwords = spacy.lang.en.stop_words.STOP_WORDS

from spacy.util import minibatch, compounding
from spacy import displacy
from spacy.tokens import Doc

from spacy.scorer import Scorer

## Load data from toxic_basic Pickle File

%%time
'''
last load time:

CPU times: user 67 ms, sys: 46.7 ms, total: 114 ms
Wall time: 114 ms
'''

# load toxic_basic pickle into dataframe
path_toxic_basic = "../data/toxic_basic.pkl"

toxic_df = pd.read_pickle(path_toxic_basic)

toxic_df.info()

# Removing tuples here...
# Convert training text and training outcomes into a list of tuples

# toxic_df["tuples"] = toxic_df.apply(lambda row: (row['comment_text'], row['toxic']), axis=1)

# toxic_df['tuples'][0]

# Simple Train Test Split

As our process should first determine whether the text is toxic or not toxic, we'll make a simplified stratified train test split, ensuring our balance of toxic and non toxic rows are proportionally distributed.

For now, we won't be too concerned with the proportion of sub-categories, as our first step will be to filter not toxic from toxic, then run parallel operations for each toxic sub-category, as toxic sub-categories are not mutually exclusive.

## Stratified Split maintaining ratio of toxic to not toxic texts


# check current columns
toxic_df.columns

# split df into X(independent) and y(depenendent) groups
ind_cols = ['comment_text', 'uppercase_proportion']

X = toxic_df[ind_cols]
y = toxic_df.drop(columns=ind_cols)

print(f"X columns: {X.columns}\ny columns:{y.columns}")

# Train Test Split. Stratified on y['toxic']
X_train, X_test, y_train, y_test = train_test_split(X, y, 
                                                    test_size=0.33, 
                                                    random_state=42, 
                                                    stratify=y['toxic'])

## Preserve Test Train Split Files

# ! ls ../data/basic_df_split/

X_train.to_pickle('../data/basic_df_split/basic_X_train.pkl')
X_test.to_pickle('../data/basic_df_split/basic_X_test.pkl')
y_train.to_pickle('../data/basic_df_split/basic_y_train.pkl')
y_test.to_pickle('../data/basic_df_split/basic_y_test.pkl')

# spaCy

Let's try out spaCy, a nlp processing library!

- https://course.spacy.io/en/chapter1
- [text classification with spaCy](https://www.dataquest.io/blog/tutorial-text-classification-in-python-using-spacy/) 
- [customized list of stopwords](https://spacy.io/usage/linguistic-features#stop-words)  
- [Split Series into list of sentences](https://pandas.pydata.org/pandas-docs/stable/reference/api/pandas.Series.str.cat.html)  
- [contractions](https://theslaps.medium.com/cant-stand-don-t-want-contractions-with-spacy-39715cac2ebb)  


- [v2.spacy.io](https://v2.spacy.io/)

# Train spaCy Model for Classification

## Establish spaCy Pipeline

"spaCy's components are supervised models for text annotations, meaning that they can only learn to reproduce examples, not guess new labels from raw text."

By default, spaCy's text categorizer is a simple convolutional neural network.

Resources:
- [for emojis](https://spacy.io/universe/project/spacymoji)  

Code is modified from tutorial here:

Resource:
https://www.machinelearningplus.com/nlp/custom-text-classification-spacy/

# ! python -m spacy download en_core_web_lg

Resources
- [spaCy docs: scorer](https://spacy.io/api/scorer)  

- [F-Score](https://en.wikipedia.org/wiki/F-score)  

## if the model is not yet locally available
# ! python -m spacy download en_core_web_lg

import en_core_web_lg
nlp = en_core_web_lg.load()

# Provide scoring pipeline
scorer = Scorer(nlp)

nlp.pipe_names

# tagger = nlp.create_pipe('tagger')
textcat = nlp.create_pipe('textcat')

# nlp.add_pipe(tagger)
nlp.add_pipe(textcat)

# textcat.add_label("TOXIC")
# textcat.add_label("SEVERE_TOXIC")
# textcat.add_label("OBSCENE")
# textcat.add_label("THREAT")
# textcat.add_label("INSULT")
# textcat.add_label("IDENTITY_HATE")
# # textcat.add_label("NOT TOXIC")

def add_labels_helper(s):
    '''
    takes dataframe or series, 
    unpacks col labels and adds each as label to textcat
        formatted as uppercase
    '''
    
    for col in s.columns:
        print(col)
        textcat.add_label(col.upper())

add_labels_helper(y_train)
    

nlp.pipe_names

textcat.labels

[zipping from dict of unknown size](https://stackoverflow.com/a/40658867)

# from spacy.tokens import Doc
# from spacy.training import Example


def txt_and_multi_cat(txt_series, multi_cat_df):
        
    # convert each series or series slice to list
    t = txt_series.tolist()

    
    # get name for each dependent column
    cats = [multi_cat_df[cat].name.upper() for cat in multi_cat_df.columns]
    print(cats)
    
    
    cat_vals = multi_cat_df.values.tolist()
    
    c = [{cats[i]: v for i, v in enumerate(row)} for row in cat_vals]
    
    
#     for row in cat_vals:
#         print(row)
# #         row_cat = dict()
        
# #         for i, v in enumerate(row):
# # #             print(i, v)
# #             row_cat[cats[i]] = v
    
#         row_cat = {cats[i]: v for i, v in enumerate(row)}
        
        
        
#         annotations_list.append(row_cat)

#     print(len(t), len(annotations_list))
#     for i in annotations_list:
#         print(i)

    
# def txt_and_cat(txt_series, cat_series):
        
#     # convert each series or series slice to list
#     t = txt_series.tolist()
#     c = cat_series.tolist()
    
#     # format categories
#     c = [{"TOXIC": bool(y), "NOT TOXIC": not bool(y)} for y in c]
    c = [{'cats': i} for i in c]
    
    docs = list(zip(t, c))
    
    return docs


tot = 20
tiny_X = X_train['comment_text'][:tot]
tiny_y = y_train[:tot]

txt_and_multi_cat(tiny_X, tiny_y)
# print(tiny_X, tiny_y)

[Article for v3](https://medium.com/analytics-vidhya/building-a-text-classifier-with-spacy-3-0-dd16e9979a)

# ! ls ../models/base_config.cfg
# ! python -m spacy init fill-config ../models/base_config.cfg config.cfg --diff

# ! python -m spacy validate

<!-- X columns: Index(['comment_text', 'uppercase_proportion'], dtype='object')
y columns:Index(['toxic', 'severe_toxic', 'obscene', 'threat', 'insult', 'identity_hate',
       'tuples'],
      dtype='object') -->

# formatting list of tuples for spacy training
train_txt = X_train['comment_text']
train_cat = y_train[['toxic', 'severe_toxic', 'obscene', 'threat', 'insult', 'identity_hate']]

train_docs = txt_and_multi_cat(train_txt, train_cat)

test_txt = X_test['comment_text']
test_cat = y_test[['toxic', 'severe_toxic', 'obscene', 'threat', 'insult', 'identity_hate']]

test_docs = txt_and_multi_cat(test_txt, test_cat)

# this should be the correct format expected by the trainer

# print(train_docs[0][1])
first_five = [i for i in train_docs[:5]]

for i in first_five:
    print(f"{i[0][:30]}, {i[1]}")

https://www.machinelearningplus.com/nlp/custom-text-classification-spacy

# Evaluate Function

The evaluate function preserves the scores for each epoch. Scores are broken down categorically, and will assist us in checking the model's effectiveness.

def evaluate(tokenizer, textcat, val_texts, val_cats, thresh=0.5):
    
    docs = (tokenizer(val_text) for val_text in val_texts)  
    
    # create dict of results
    evals_by_cat = dict()
    
    for i, doc in enumerate(textcat.pipe(docs)):
        
        # actual scores 
        gold = val_cats[i]['cats']
        
        for label, score in doc.cats.items():
            
            # add label to dict if not already present
            if label not in evals_by_cat:
                evals_by_cat[label] = {'tp':0,
                                       'fp':0,
                                       'fn':0,
                                       'tn':0,}
            
            if score >= thresh and gold[label] >= thresh:
                evals_by_cat[label]['tp'] += 1

            elif score >= thresh and gold[label] < thresh:
                evals_by_cat[label]['fp'] += 1

            elif score < thresh and gold[label] < thresh:
                evals_by_cat[label]['tn'] += 1
            
            elif score < thresh and gold[label] >= thresh:
                evals_by_cat[label]['fn'] += 1
    
    for key in evals_by_cat.keys():

        # local scope variables to ease reading & debugging
        tp = evals_by_cat[key]['tp']
        fp = evals_by_cat[key]['fp']
        fn = evals_by_cat[key]['fn']
        tn = evals_by_cat[key]['tn']
        
        # precision
        # edge case: avoid dividing by zero: precision = 1 when fp = 0
        if tp + fp == 0:
            evals_by_cat[key]['precision'] = 1
        else:    
            evals_by_cat[key]['precision'] = tp / (tp + fp)
        
        # recall
        # edge case: avoid dividing by zero: recall = 1 when fn = 0
        if tp + fn == 0:
            evals_by_cat[key]['recall'] = 1
        else:    
            evals_by_cat[key]['recall'] = tp / (tp + fn)
            
        # local variables to ease reading & debugging
        precision = evals_by_cat[key]['precision']
        recall = evals_by_cat[key]['recall']
        
        # f score
        if precision  + recall == 0:
            evals_by_cat[key]['f_score'] = 0.0
        else:
            evals_by_cat[key]['f_score'] = 2 * (precision * recall) / (precision + recall)
    
    # preserve spaCy's native losses metric
    evals_by_cat['TEXTCAT_LOSSES'] = losses['textcat']

    return evals_by_cat

# variable used in testing
# size = 1000
train_data = train_docs
dev_texts = [i[0] for i in test_docs]
dev_cats = [i[1] for i in test_docs]

# Do 5 Training Iterations 

We'll begin with 5 training iterations to help us estimate how long training will be and get a few baseline predictions.

Each iteration takes roughly 12 minutes, so 5 will be about an hour

## log from previous run:
Training the model.
Iterations: 5
Timer Begins:14924.26
epoch 1 start time: 14924.26
textcat losses: 3.300574591791758
epoch 1 end: 15645.511968644, epoch elapsed: 721.25

epoch 2 start time: 15645.52
textcat losses: 3.2344222257217226
epoch 2 end: 16362.236106837, epoch elapsed: 716.72

epoch 3 start time: 16362.24
textcat losses: 2.8232882740304106
epoch 3 end: 17072.387902298, epoch elapsed: 710.15

epoch 4 start time: 17072.39
textcat losses: 2.6497837365776666
epoch 4 end: 17787.474139616, epoch elapsed: 715.08

epoch 5 start time: 17787.48
textcat losses: 2.3880158715215387
epoch 5 end: 18499.152896641, epoch elapsed: 711.68

Training complete. End:18499.16 Training Elapsed: 3574.90

CPU times: user 1h 39min 33s, sys: 1h 3min 52s, total: 2h 43min 26s
Wall time: 59min 34s

%%time
from spacy.util import minibatch, compounding

log_list = list()

# start = timer()
# # ...
# end = timer()
# print(end - start) # Time in seconds, e.g. 5.38091952400282

#("Number of training iterations", "n", int))
n_iter=5

# Disabling other components
other_pipes = [pipe for pipe in nlp.pipe_names if pipe != 'textcat']
with nlp.disable_pipes(*other_pipes):  
    optimizer = nlp.begin_training()

    start_cumulative = timer()
    print(f"Training the model.\nIterations: {n_iter}\nTimer Begins:{start_cumulative:.2f}")
#     print('{:^5}\t{:^5}\t{:^5}\t{:^5}'.format('LOSS', 'P', 'R', 'F'))

    # Performing training
    for i in range(n_iter):
        start_epoch = timer()
        print(f'epoch {i + 1} start time: {start_epoch:.2f}')
        losses = {}
        batches = minibatch(train_data, size=compounding(4., 32., 1.001))
        
        for batch in batches:
#             print(f"new batch at :{timer():.2f}")
            texts, annotations = zip(*batch)
            nlp.update(texts, 
                       annotations, 
                       sgd=optimizer, 
                       drop=0.2,
                       losses=losses)

      # Calling the evaluate() function and printing the scores
        with textcat.model.use_params(optimizer.averages):
            evals_by_cat = evaluate(nlp.tokenizer, textcat, dev_texts, dev_cats)
            # reduced scores for testing
            print(f'textcat losses:', losses['textcat'])
            log_list.append(evals_by_cat)
            
            end_epoch = timer()
            time.strftime("%Hh%Mm%Ss", time.gmtime(4*3600+13*60+6)) 
            print(f'epoch {i + 1} end: {end_epoch}, epoch elapsed: {end_epoch-start_epoch:.2f}\n')
         
            
#             for key in evals_by_cat.keys():
#                 print(f'{key}:\n{evals_by_cat[key]}')
#         print('{0:.3f}\t{1:.3f}\t{2:.3f}\t{3:.3f}'  
#               .format(losses['textcat'], scores['textcat_p'],
#                       scores['textcat_r'], scores['textcat_f']))

end_cumulative = timer()
print(f"Training complete. End:{end_cumulative:.2f} Training Elapsed: {end_cumulative - start_cumulative:.2f}\n")

# len(log_list)

import pickle

with open('../logs/log_list_2.pkl', 'wb') as f:
    pickle.dump(log_list, f)

# log_list

# with open('../logs/log_list_1.pkl', 'rb') as f:
#     test_list = pickle.load(f)

test_list

first_df = pd.DataFrame.from_dict(evals_by_cat['INSULT'], orient='index', columns=['INSULT_epoch1'])
second_df = pd.DataFrame.from_dict(evals_by_cat['TOXIC'], orient='index', columns=['TOXIC_epoch1'])


pd.concat([first_df, second_df], axis=1)

evals_df = pd.DataFrame.from_dict(evals_by_cat)


evals_df

one_text = dev_texts[11]
# one_text

print(one_text)

doc=nlp(one_text)
doc.cats 

# sentiment_sum = sum([i.sentiment for i in doc])
# print(sentiment_sum)
# doc[0].vector

def doc_check(tok):
    '''
    helper function for getting only desired lemmas from spaCy doc
    argument: doc.token
    
    checks for rejection conditions
        not alpha
        pronoun
        stopword
        
    returns True if no rejection conditions are met
    
    ''' 
    # reject if not alpha
    if tok.is_alpha == False:
        return False
    
    # reject if pronoun
    if tok.lemma_ == "-PRON-":
        return False
    
    # reject if stopword
    if tok.is_stop == True:
        return False

    # if not rejected, return true
    return True

lemmas_lc = [i.lemma_.lower() for i in doc if doc_check(i)]
lemmas_lc

# sentiment_sum = sum([i.sentiment for i in doc])
# print(sentiment_sum)

[expanding contractions](https://gist.github.com/widiger-anna/deefac010da426911381c118a97fc23f) 
[contractions](https://theslaps.medium.com/cant-stand-don-t-want-contractions-with-spacy-39715cac2ebb)  


[text wrangling](https://www.kdnuggets.com/2018/08/practitioners-guide-processing-understanding-text-2.html)  


[nlp nltk vs spacy](https://www.activestate.com/blog/natural-language-processing-nltk-vs-spacy/)  

[pytorch](https://pytorch.org/https://pytorch.org/)  

[text classification in python with spacy (try this one!)](https://www.dataquest.io/blog/tutorial-text-classification-in-python-using-spacy/https://www.dataquest.io/blog/tutorial-text-classification-in-python-using-spacy/)  

https://towardsdatascience.com/machine-learning-nlp-text-classification-using-scikit-learn-python-and-nltk-c52b92a7c73a

! ls ../models
nlp.to_disk("../models/spacy_multi_cat_model/")
! ls ../models

! ls

! ls ../models

# Do 20 Additional Training Iterations

%%time
log_list = list()

# start = timer()
# # ...
# end = timer()
# print(end - start) # Time in seconds, e.g. 5.38091952400282

#("Number of training iterations", "n", int))
n_iter=20

# Disabling other components
other_pipes = [pipe for pipe in nlp.pipe_names if pipe != 'textcat']
with nlp.disable_pipes(*other_pipes):  
    optimizer = nlp.begin_training()

    start_cumulative = timer()
    print(f"Training the model.\nIterations: {n_iter}\nTimer Begins:{start_cumulative:.2f}")
#     print('{:^5}\t{:^5}\t{:^5}\t{:^5}'.format('LOSS', 'P', 'R', 'F'))

    # Performing training
    for i in range(n_iter):
        start_epoch = timer()
        print(f'epoch {i + 1} start time: {start_epoch:.2f}')
        losses = {}
        batches = minibatch(train_data, size=compounding(4., 32., 1.001))
        
        for batch in batches:
#             print(f"new batch at :{timer():.2f}")
            texts, annotations = zip(*batch)
            nlp.update(texts, 
                       annotations, 
                       sgd=optimizer, 
                       drop=0.2,
                       losses=losses)

      # Calling the evaluate() function and printing the scores
        with textcat.model.use_params(optimizer.averages):
            evals_by_cat = evaluate(nlp.tokenizer, textcat, dev_texts, dev_cats)
            # reduced scores for testing
            print(f'textcat losses:', losses['textcat'])
            log_list.append(evals_by_cat)
            
            end_epoch = timer()
            time.strftime("%Hh%Mm%Ss", time.gmtime(4*3600+13*60+6)) 
            print(f'epoch {i + 1} end: {end_epoch}, epoch elapsed: {end_epoch-start_epoch:.2f}\n')
         
            
#             for key in evals_by_cat.keys():
#                 print(f'{key}:\n{evals_by_cat[key]}')
#         print('{0:.3f}\t{1:.3f}\t{2:.3f}\t{3:.3f}'  
#               .format(losses['textcat'], scores['textcat_p'],
#                       scores['textcat_r'], scores['textcat_f']))

end_cumulative = timer()
print(f"Training complete. End:{end_cumulative:.2f} Training Elapsed: {end_cumulative - start_cumulative:.2f}\n")

## Preserve log list

These logs are currently saved as dictionaries. We'll preserve them so we can

with open('../logs/log_list_2.pkl', 'wb') as f:
    pickle.dump(log_list, f)

## Preserve Model

! ls ../models
nlp.to_disk("../models/spacy_multi_cat_model/")
! ls ../models

# I left off here!!!

https://v2.spacy.io/usage/processing-pipelines#pipelines
https://v2.spacy.io/usage/processing-pipelines
